+++ src/jiratui/widgets/commons/widgets.py
@@ -79,7 +79,9 @@ from textual.events import Key
 from textual.message import Message
 from textual.validation import Number, ValidationResult
 from textual.widgets import (
+    Collapsible,
     Input,
+    Link,
     MaskedInput,
     Select,
     SelectionList,
@@ -92,6 +94,7 @@ from textual.widgets.selection_list import Selection
 
 from jiratui.actions.constants import SupportedActions
 from jiratui.actions.keys import get_application_key_bindings
+from jiratui.utils.adf import extract_web_links_from_markdown
 from jiratui.utils.ui_actions import Actionable, UIAction
 from jiratui.widgets.base import DateInput
 from jiratui.widgets.commons.base import (
@@ -1624,6 +1627,9 @@ class ReadOnlyPlainTextTextAreaWidget(TextArea):
 
     Use it in the Info tab or to display comment's content or to display fields in the read-only work item details
     screen.
+
+    This widget provides a method to extract web links found in the text content being displayed. This is useful to help
+    the user navigate the links in the text; Markdown widgets do not support this.
     """
 
     def __init__(
@@ -1683,6 +1689,14 @@ class ReadOnlyPlainTextTextAreaWidget(TextArea):
     @property
     def field_title(self) -> str:
         return self.__title
+
+    def extract_web_links(self) -> list[Link]:
+        if not self.text:
+            return []
+        web_links_in_text: list[dict] = extract_web_links_from_markdown(self.text)
+        return [
+            Link(text=url.get('title', 'link'), url=url.get('url')) for url in web_links_in_text
+        ]
 
 
 class EmptyTextAreaStaticWidget(Static):
@@ -2382,3 +2396,14 @@ class ActionableTabbedContent(Actionable, TabbedContent, inherit_bindings=False)
         tabs = self.query_one(Tabs)
         if tabs.has_focus:
             tabs.action_previous_tab()
+
+
+class WebLinksCollapsible(Collapsible):
+    """A Collapsible to display web links.
+
+    This is useful for displaying links found in text content, e.g. comments, description, and other textarea (custom)
+    fields.
+    """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, title='View links found in the text', **kwargs)