+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+
+- Extract and display web links found in Markdown context. If links are found in the content then a collapsible section
+will appear in the Info tab's text panes (description, environment, etc.) so the use can open the desired
+links. By [@whyisdifficult](https://github.com/whyisdifficult) in https://github.com/whyisdifficult/jiratui/pull/347
+
 ## [1.13.1] 2026-08-28
 
 ### Added